fix shutting down issue
When launched in a terminal:
1. Waybar does not quit immediately when C-c is pressed.
2. Waybar will crash if C-c again.

When SIGINT is received, waybat tries to quit, and gtk isses an `_IO_flush_all`, which needs to
accquire io lock first via `__lll_lock_wait_private`.

But this lock is currently owned by ipc message receiver which is blocked by fgets(). It does not
return unless something is read. In this case, if Hyprland sends some messages, then fgets turns, then
lock is released, and main thread quits normally.

But if nothing is send, quit then hangs.